Changes file structure to use tab routing using tab package from expo router
